Final changes
Swapped out old uart code so that it will actually send bytes of data
instead of strings. This helps accuracy a lot because the data being
sent isn't being rounded at all between the two systems.

TODO: test if cliff and edge detection still works well after change.
Should work fine, but should probably test it before lab.

+++ structs.h
@@ -16,9 +16,9 @@ typedef struct Scan {
 typedef struct TallObject {
     int obj_num;
     int angle;
-    float dist;
+    double dist;
     int radial_width;
-    float linear_width;
+    double linear_width;
 } TallObject;
 
 #endif /* STRUCTS_H_ */
